@@ -10,17 +10,24 @@
   },
   "keywords": [],
   "author": "",
-  "license": "ISC",
+  "license": "MIT",
   "devDependencies": {
+    "@types/jquery": "^2.0.40",
+    "@types/lodash": "^4.14.52",
+    "@types/minimist": "^1.2.0",
+    "@types/moment": "^2.13.0",
+    "@types/react": "^15.0.12",
+    "@types/react-dom": "^0.14.23",
     "ts-loader": "^0.8.2",
-    "typescript": "^2.0.0-dev.20160628",
-    "typings": "^1.3.1",
+    "typescript": "^2.2.1",
     "webpack": "^1.13.1"
   },
   "dependencies": {
+    "jquery": "^3.1.1",
     "lodash": "^4.17.4",
-    "moment": "^2.13.0",
-    "react": "^15.1.0",
-    "react-dom": "^15.1.0"
+    "minimist": "^1.2.0",
+    "moment": "^2.17.1",
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2"
   }
 }
